place, approval
권한 수정

+++ WorkFlow_Project_BE/src/main/java/com/kh/workflow/place/controller/PlaceController.java
@@ -29,6 +29,7 @@ import com.kh.workflow.place.model.service.PlaceService;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
+import org.springframework.security.core.Authentication;
 
 @RestController
 @RequestMapping("/place")
@@ -43,7 +44,15 @@ public class PlaceController {
     public ResponseEntity<String> insertPlace(
             @RequestPart("place") Hub h,
             @RequestPart(value = "file", required = false) MultipartFile file,
-            HttpSession session) {
+            HttpSession session,
+            Authentication authentication) {
+    	
+    	if(!authentication.getAuthorities().stream()
+    			.anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+    		return ResponseEntity
+    				.status(HttpStatus.FORBIDDEN)
+    				.body("관리자만 장소를 등록할 수 있습니다.");
+    	}
 
         // 기본 상태 설정
         if (h.getHubStatus() == null || h.getHubStatus().isBlank()) {
@@ -144,7 +153,16 @@ public class PlaceController {
             @PathVariable int hubNo,
             @RequestPart("place") Hub h,
             @RequestPart(value = "file", required = false) MultipartFile file,
-            HttpSession session) {
+            HttpSession session,
+            Authentication authentication) {
+    	
+    	if(!authentication.getAuthorities().stream()
+    			.anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+    		
+    		return ResponseEntity
+    				.status(HttpStatus.FORBIDDEN)
+    				.body("관리자만 장소를 수정할 수 있습니다.");
+    	}
 
         h.setHubNo(hubNo);
 
@@ -186,8 +204,17 @@ public class PlaceController {
     // 장소 종료
     @DeleteMapping("/{hubNo}")
     public ResponseEntity<String> deletePlace(
-            @PathVariable int hubNo) {
+            @PathVariable int hubNo,
+            Authentication authentication) {
 
+    	if(!authentication.getAuthorities().stream()
+    			.anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"))) {
+    		
+    		return ResponseEntity
+    				.status(HttpStatus.FORBIDDEN)
+    				.body("관리자만 장소를 삭제할 수 있습니다.");
+    	}
+    	
         int result =
                 placeService.deletePlace(hubNo);
